% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hear.R
\name{Hear}
\alias{Hear}
\title{hear no evil}
\usage{
Hear(x)
}
\arguments{
\item{x}{an input}
}
\value{
The string \code{"No Evil."}.
}
\description{
A function that will only "hear" an input containing variants on the string `"no evil"`.
`Hear` is not case sensitive and will remove punctuation so `"no_Evil"` and `"nO- _eVil"` will also be heard.
`Hear` is also sensitive to value names so passing an object `no_evil <- "asdf"` containing any arbitrary strings will also be heard.
Anything that is not specifically `"no evil"` or some acceptable variant is assumed to be evil and will not be heard.
}
\examples{
Hear(1)
Hear("No Evil")
no_evil <- "good stuff"
Hear(no_evil)
evil <- "good_stuff"
Hear(evil)
Hear(NA)
}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
