% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speak.R
\name{Speak}
\alias{Speak}
\title{speak no evil}
\usage{
Speak(...)
}
\arguments{
\item{...}{an input, doesn't matter, isn't used}
}
\value{
Plays an audio clip of the words "No Evil.".
}
\description{
Plays an audioclip of the words "No evil".
On Windows and Linux machines the function should directly call the system default player of .wav files.
On Mac machines it might be nessisary to set the default with something like tuneR::setWavPlayer("/Applications/'QuickTime Player.app'/Contents/MacOS/'QuickTime Player'").
}
\examples{
\dontrun{
Speak()
}

}
\author{
Greg Pilgrim \email{gpilgrim2670@gmail.com}
}
