% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hear.R
\name{Hear}
\alias{Hear}
\title{hear no evil}
\usage{
Hear(x)
}
\arguments{
\item{x}{an input}
}
\value{
The string \code{"No Evil."}.
}
\description{
hear no evil
}
\examples{
Hear(1)
Hear("No Evil")
no_evil <- "good stuff"
Hear(no_evil)
evil <- "good_stuff"
Hear(evil)
}
